requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
DateTime==5.5
distlib==0.3.9
dj-database-url==3.0.1
Django==5.1.1
django-environ==0.12.0
django-extensions==4.1
django-request==1.6.3
django-widget-tweaks==1.5.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
filelock==3.16.1
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
packaging==24.2
pillow==11.3.0
pipenv==2024.4.0
platformdirs==4.3.6
PyJWT==2.10.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.3
requests==2.32.3
setuptools==75.1.0
six==1.16.0
sqlparse==0.5.1
tzdata==2024.2
uritemplate==4.2.0
urllib3==2.2.3
virtualenv==20.28.0
whitenoise==6.11.0
zope.interface==7.0.3
